+++ gpytorch/lazy/half_non_lazy_tensor.py
@@ -10,10 +10,10 @@ class HalfNonLazyTensor(NonLazyTensor):
         return super()._getitem(*indices)
 
     def _matmul(self, rhs):
-        return super()._matmul(rhs.half()).float()
+        return super()._matmul(rhs.half().float()).float()
 
     def _t_matmul(self, rhs):
-        return super()._t_matmul(rhs.half()).float()
+        return super()._t_matmul(rhs.half().float()).float()
 
     def _quad_form_derivative(self, left_vecs, right_vecs):
         return (res.float() for res in super()._quad_form_derivative(left_vecs, right_vecs))
